fix: Address clippy warnings and formatting issues
Fixes for CI linting:
- Collapsed nested if statements in WAL flush thread (clippy::collapsible_if)
- Fixed benchmark access to DurabilityMode::Async (use struct syntax)
- Added missing WriteOps trait import in gallifreydb benchmark
- Removed unnecessary cast in temporal_vector benchmark
- Auto-formatted test file changes

All 648 tests pass.
Clippy passes with --all-targets --all-features -D warnings.

+++ benches/temporal_vector.rs
@@ -461,7 +461,7 @@ fn bench_semantic_evolution_memory_overhead(c: &mut Criterion) {
                 let all_snapshots_vectors: Vec<Vec<Vec<f32>>> = (0..10)
                     .map(|snapshot_idx| {
                         (0..vector_count)
-                            .map(|i| gen_vector(384, (snapshot_idx + i) as usize))
+                            .map(|i| gen_vector(384, snapshot_idx + i))
                             .collect()
                     })
                     .collect();
